Update dota-2.mdx
Dear Overwolf,

Your links to the hero IDs are very much outdated. The data on the GitHub repo is 9 years old and 13 heroes are missing.

I suggest using the Dota Coach website, as it is fully updated, provides all heroes and on top of the JSON file also allows for on-line analysis and has much more information about the Dota 2 heroes.

What do you think?